requirements.txt
pick==0.6.4
python-chess==0.27.3